\docType{data}
\name{pattern}
\alias{pattern}
\title{Summary of fixation pattern generated by get_pattern() function}
\format{The data set aggregates the fixation probability data, and show
  the general fixation pattern. The data frame has the following core
  columns:
    \describe{
      \item{\code{condition}}{the id of the condition}
      \item{\code{region}}{the id of the region}
      \item{\code{0}}{averaged fixation probability at time 0}
      \item{\code{2500}}{averaged fixation probability at time 0}}}
\usage{
  data(pattern)
}
\description{
  This data set is generated by get_pattern() function, and describes the 
  general fixation pattern for different conditions. The values under time
  variables are the aggregated fixation probabilities for each spatio-temporal
  unit. Users may use the data set to make plots of the pattern by themselves
  or use the dataset as argument of plot_pattern() and lineplot() functions.
}
\keyword{datasets}
